third_party/WebKit/LayoutTests/http/tests/notifications/serviceworkerregistration-document-click.html

<!doctype html>
<html>
  <head>
    <title>Notifications: ServiceWorkerRegistration.showNotification().</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
    <script src="resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that the showNotification() function resolves a promise, and that the
      // notificationclick event gets fired on the Service Worker when we simulate a
      // click on it. This test requires the test runner.

      async_test(function(test) {
          var scope = 'resources/scope/serviceworkerregistration-document-click',
              script = 'resources/instrumentation-service-worker.js';

          testRunner.setPermission('notifications', 'granted', location.origin, location.origin);

          var workerInfo = null;
          getActiveServiceWorkerWithMessagePort(test, script, scope).then(function(info) {
              workerInfo = info;

              // (1) Display a Web Notification from the document.
              assert_inherits(workerInfo.registration, 'showNotification', 'showNotification() must be exposed.');
              return workerInfo.registration.showNotification(scope, {
                  body: 'Hello, world!',
                  icon: '/icon.png'
              });
          }).then(function() {
              // (2) Simulate a click on the notification that has been displayed.
              testRunner.simulateWebNotificationClick(scope, -1 /* action_index */);

              workerInfo.port.addEventListener('message', function(event) {
                  if (typeof event.data != 'object' || !event.data.command) {
                      assert_unreached('Received an invalid message from the Service Worker.');
                      return;
                  }

                  // (3) Verify that the click event was received by the Service Worker.
                  assert_equals(event.data.command, 'click');
                  assert_equals(event.data.notification.title, scope);

                  test.done();
              });
          }).catch(unreached_rejection(test));

      }, 'Clicking on a notification displayed through showNotification() fires a Service Worker event.');
    </script>
  </body>
</html>
